requirements.txt
amqp==2.5.0
aniso8601==7.0.0
attrs==19.1.0
billiard==3.6.0.0
celery==4.3.0
Click==7.0
Flask==1.0.3
flask-restplus==0.12.1
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
kombu==4.6.3
MarkupSafe==1.1.1
pyrsistent==0.15.2
python-dateutil==2.8.0
pytz==2019.1
redis==3.2.1
six==1.12.0
vine==1.3.0
Werkzeug==0.15.4